moving webpack configs to separated directory
