+++ config/webpack.dev.config.js
@@ -1,15 +1,15 @@
 var webpack = require('webpack');
 var path    = require('path');
-var config  = require('./webpack.config');
+var config  = require('../webpack.config');
 
 config.output = {
   filename: '[name].bundle.js',
-  path: path.resolve(__dirname, 'client')
+  path: path.resolve(__dirname, '..', 'client')
 };
 
 config.devServer = {
   port: 3000,
-  contentBase: './client',
+  contentBase: '../client',
   hot: true,
   stats: { colors: true },
   inline: true,
