+++ config/webpack.prod.config.js
@@ -1,11 +1,11 @@
 var webpack = require('webpack');
 var path    = require('path');
-var config  = require('./webpack.config');
+var config  = require('../webpack.config');
 
 config.output = {
   filename: '[name].bundle.js',
   publicPath: '',
-  path: path.resolve(__dirname, 'dist')
+  path: path.resolve(__dirname, '..', 'dist')
 };
 
 config.plugins = config.plugins.concat([
